🐞 fix(Network): Updated Network
Made a heartbet sender so that heartbeats gets sent out periodically. In addition made function that send updates on all orders, and updates on states from the slaves. These have to be placed somewhere where it makes sense to put them. Ack-msg framework is up, just have to add a case in the reciever to send an ack message back for x amount of requests.